modules/web/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
jinja2>=3.1
pyyaml>=6.0
ruamel.yaml>=0.18
aiofiles>=23.1
